requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.11
pydantic==2.12.2
pydantic-settings==2.13.1
python-dotenv==1.2.2
jinja2==3.1.6
weasyprint==68.0
pydyf==0.11.0
pypdf==6.10.2
reportlab==4.4.10
python-multipart==0.0.26
ofxparse>=0.21
python-dateutil>=2.9.0
